+++ src/components/FloatingTip/FloatingTip.tsx
@@ -5,7 +5,7 @@ import { ModalCloseReason } from "../Modal";
 import Popover, { PopoverProps } from "../Popover";
 import * as Styled from "./styled";
 import { Placement } from "@floating-ui/react";
-import { AutoPlacement } from "../../utils/placement";
+import { AutoPlacement } from "../../hooks/usePlacement";
 
 export type FloatingTipCloseReason = "clickCloseIcon";
